# core/generated/swagger-ui/schema-registry-api-spec.yaml
...
swagger: "2.0"
info:
  version: "v1"
  title: "Confluent Schema Registry"
schemes:
- "http"
- "https"
paths:
...
  /compatibility/subjects/{subject}/versions/{version}:
    post:
      summary: "Test input schema against a particular version of a subject's schema\
        \ for compatibility."
      description: "the compatibility level applied for the check is the configured\
        \ compatibility level for the subject (http:get:: /config/(string: subject)).\
        \ If this subject's compatibility level was never changed, then the global\
        \ compatibility level applies (http:get:: /config)."
      operationId: "testCompatibilityBySubjectName"
      consumes:
      - "application/vnd.schemaregistry.v1+json"
      - "application/vnd.schemaregistry+json"
      - "application/json"
      - "application/octet-stream"
      produces:
      - "application/vnd.schemaregistry.v1+json"
      - "application/vnd.schemaregistry+json; qs=0.9"
      - "application/json; qs=0.5"
      parameters:
      - name: "Content-Type"
        in: "header"
        required: false
        type: "string"
      - name: "Accept"
        in: "header"
        required: false
        type: "string"
      - name: "subject"
        in: "path"
        description: "Subject of the schema version against which compatibility is\
          \ to be tested"
        required: true
        type: "string"
        pattern: ".+"
      - name: "version"
        in: "path"
        description: "Version of the subject's schema against which compatibility\
          \ is to be tested. Valid values for versionId are between [1,2^31-1] or\
          \ the string \"latest\".\"latest\" checks compatibility of the input schema\
          \ with the last registered schema under the specified subject"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        description: "Schema"
        required: true
        schema:
          $ref: "#/definitions/RegisterSchemaRequest"
      - name: "Authorization"
        in: "header"
        required: false
        type: "string"
      - name: "Cookie"
        in: "header"
        required: false
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/CompatibilityCheckResponse"
        404:
          description: "Error code 40401 -- Subject not found\nError code 40402 --\
            \ Version not found"
        422:
          description: "Error code 42201 -- Invalid schema or schema type\nError code\
            \ 42202 -- Invalid version"
        500:
          description: "Error code 50001 -- Error in the backend data store"
...
  /schemas/ids/{id}:
    get:
      summary: "Get the schema string identified by the input ID."
      description: ""
      operationId: "getSchema"
      consumes:
      - "application/vnd.schemaregistry.v1+json"
      - "application/vnd.schemaregistry+json"
      - "application/json"
      - "application/octet-stream"
      produces:
      - "application/vnd.schemaregistry.v1+json"
      - "application/vnd.schemaregistry+json; qs=0.9"
      - "application/json; qs=0.5"
      parameters:
      - name: "id"
        in: "path"
        description: "Globally unique identifier of the schema"
        required: true
        type: "integer"
        format: "int32"
      - name: "format"
        in: "query"
        required: false
        type: "string"
      - name: "fetchMaxId"
        in: "query"
        required: false
        type: "boolean"
        default: false
      - name: "Authorization"
        in: "header"
        required: false
        type: "string"
      - name: "Cookie"
        in: "header"
        required: false
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/SchemaString"
        404:
          description: "Error code 40403 -- Schema not found\n"
        500:
          description: "Error code 50001 -- Error in the backend data store\n"
...
  /subjects/{subject}:
    post:
      summary: "Check if a schema has already been registered under the specified\
        \ subject. If so, this returns the schema string along with its globally unique\
        \ identifier, its version under this subject and the subject name."
      description: ""
      operationId: "lookUpSchemaUnderSubject"
      consumes:
      - "application/vnd.schemaregistry.v1+json"
      - "application/vnd.schemaregistry+json"
      - "application/json"
      - "application/octet-stream"
      produces:
      - "application/vnd.schemaregistry.v1+json"
      - "application/vnd.schemaregistry+json; qs=0.9"
      - "application/json; qs=0.5"
      parameters:
      - name: "subject"
        in: "path"
        description: "Subject under which the schema will be registered"
        required: true
        type: "string"
        pattern: ".+"
      - name: "deleted"
        in: "query"
        required: false
        type: "boolean"
      - in: "body"
        name: "body"
        description: "Schema"
        required: true
        schema:
          $ref: "#/definitions/RegisterSchemaRequest"
      - name: "Authorization"
        in: "header"
        required: false
        type: "string"
      - name: "Cookie"
        in: "header"
        required: false
        type: "string"
      responses:
        404:
          description: "Error code 40401 -- Subject not found\nError code 40403 --\
            \ Schema not found"
        500:
          description: "Internal server error"
          schema:
            $ref: "#/definitions/Schema"
    delete:
      summary: "Deletes the specified subject and its associated compatibility level\
        \ if registered. It is recommended to use this API only when a topic needs\
        \ to be recycled or in development environment."
      description: ""
      operationId: "deleteSubject"
      consumes:
      - "application/vnd.schemaregistry.v1+json"
      - "application/vnd.schemaregistry+json"
      - "application/json"
      - "application/octet-stream"
      produces:
      - "application/vnd.schemaregistry.v1+json"
      - "application/vnd.schemaregistry+json; qs=0.9"
      - "application/json; qs=0.5"
      parameters:
      - name: "subject"
        in: "path"
        description: "the name of the subject"
        required: true
        type: "string"
        pattern: ".+"
      - name: "permanent"
        in: "query"
        required: false
        type: "boolean"
      - name: "Authorization"
        in: "header"
        required: false
        type: "string"
      - name: "Cookie"
        in: "header"
        required: false
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              type: "integer"
              format: "int32"
        404:
          description: "Error code 40401 -- Subject not found"
        500:
          description: "Error code 50001 -- Error in the backend datastore"
  /subjects/{subject}/versions:
    get:
      summary: "Get a list of versions registered under the specified subject."
      description: ""
      operationId: "listVersions"
      consumes:
      - "application/vnd.schemaregistry.v1+json"
      - "application/vnd.schemaregistry+json"
      - "application/json"
      - "application/octet-stream"
      produces:
      - "application/vnd.schemaregistry.v1+json"
      - "application/vnd.schemaregistry+json; qs=0.9"
      - "application/json; qs=0.5"
      parameters:
      - name: "subject"
        in: "path"
        description: "Name of the Subject"
        required: true
        type: "string"
        pattern: ".+"
      - name: "deleted"
        in: "query"
        required: false
        type: "boolean"
      - name: "Authorization"
        in: "header"
        required: false
        type: "string"
      - name: "Cookie"
        in: "header"
        required: false
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              type: "integer"
              format: "int32"
        404:
          description: "Error code 40401 -- Subject not found"
        500:
          description: "Error code 50001 -- Error in the backend data store"
    post:
      summary: "Register a new schema under the specified subject. If successfully\
        \ registered, this returns the unique identifier of this schema in the registry.\
        \ The returned identifier should be used to retrieve this schema from the\
        \ schemas resource and is different from the schema's version which is associated\
        \ with the subject. If the same schema is registered under a different subject,\
        \ the same identifier will be returned. However, the version of the schema\
        \ may be different under different subjects.\nA schema should be compatible\
        \ with the previously registered schema or schemas (if there are any) as per\
        \ the configured compatibility level. The configured compatibility level can\
        \ be obtained by issuing a GET http:get:: /config/(string: subject). If that\
        \ returns null, then GET http:get:: /config\nWhen there are multiple instances\
        \ of Schema Registry running in the same cluster, the schema registration\
        \ request will be forwarded to one of the instances designated as the primary.\
        \ If the primary is not available, the client will get an error code indicating\
        \ that the forwarding has failed."
      description: ""
      operationId: "register"
      consumes:
      - "application/vnd.schemaregistry.v1+json"
      - "application/vnd.schemaregistry+json"
      - "application/json"
      - "application/octet-stream"
      produces:
      - "application/vnd.schemaregistry.v1+json"
      - "application/vnd.schemaregistry+json; qs=0.9"
      - "application/json; qs=0.5"
      parameters:
      - name: "subject"
        in: "path"
        description: "Name of the Subject"
        required: true
        type: "string"
        pattern: ".+"
      - in: "body"
        name: "body"
        description: "Schema"
        required: true
        schema:
          $ref: "#/definitions/RegisterSchemaRequest"
      - name: "Authorization"
        in: "header"
        required: false
        type: "string"
      - name: "Cookie"
        in: "header"
        required: false
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/RegisterSchemaResponse"
        409:
          description: "Incompatible schema"
        422:
          description: "Error code 42201 -- Invalid schema or schema type"
        500:
          description: "Error code 50001 -- Error in the backend data store\nError\
            \ code 50002 -- Operation timed out\nError code 50003 -- Error while forwarding\
            \ the request to the primary"
...
definitions:
  CompatibilityCheckResponse:
    type: "object"
    properties:
      is_compatible:
        type: "boolean"
...
  RegisterSchemaRequest:
    type: "object"
    properties:
      version:
        type: "integer"
        format: "int32"
      id:
        type: "integer"
        format: "int32"
      schemaType:
        type: "string"
      references:
        type: "array"
        items:
          $ref: "#/definitions/SchemaReference"
      schema:
        type: "string"
  RegisterSchemaResponse:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int32"
  Schema:
    type: "object"
    properties:
      subject:
        type: "string"
      version:
        type: "integer"
        format: "int32"
      id:
        type: "integer"
        format: "int32"
      schemaType:
        type: "string"
      references:
        type: "array"
        items:
          $ref: "#/definitions/SchemaReference"
      schema:
        type: "string"
  SchemaReference:
    type: "object"
    properties:
      name:
        type: "string"
      subject:
        type: "string"
      version:
        type: "integer"
        format: "int32"
  SchemaString:
    type: "object"
    properties:
      schemaType:
        type: "string"
        description: "Schema type"
      schema:
        type: "string"
        description: "Schema string identified by the ID"
      references:
        type: "array"
        description: "Schema references"
        items:
          $ref: "#/definitions/SchemaReference"
      maxId:
        type: "integer"
        format: "int32"
        description: "Maximum ID"
...